Structure Management
Break my code into a proper file structure for easier management
